+++ prism/src/mtbdd/PM_NondetBoundedUntil.cc
@@ -131,7 +131,7 @@ jboolean min		// min or max probabilities (true = min, false = max)
 		
 		// print occasional status update
 		if ((util_cpu_time() - start3) > UPDATE_DELAY) {
-			PM_PrintToMainLog(env, "Iteration %d (of %d): ", iters, bound);
+			PM_PrintToMainLog(env, "Iteration %d (of %d): ", iters, (int)bound);
 			PM_PrintToMainLog(env, "sol=%d nodes", DD_GetNumNodes(ddman, sol));
 			// NB: but tmp was probably bigger than sol (pre min/max-abstract)
 			PM_PrintToMainLog(env, "%.2f sec so far\n", ((double)(util_cpu_time() - start2)/1000));
